% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{AbstractStorage}
\alias{AbstractStorage}
\title{Binding of JS localStorage}
\description{
Binding of JS localStorage

Binding of JS localStorage
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractStorage-new}{\code{AbstractStorage$new()}}
\item \href{#method-AbstractStorage-setItem}{\code{AbstractStorage$setItem()}}
\item \href{#method-AbstractStorage-getItem}{\code{AbstractStorage$getItem()}}
\item \href{#method-AbstractStorage-removeItem}{\code{AbstractStorage$removeItem()}}
\item \href{#method-AbstractStorage-clone}{\code{AbstractStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractStorage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractStorage$new(input, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractStorage-setItem"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractStorage-setItem}{}}}
\subsection{Method \code{setItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractStorage$setItem(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractStorage-getItem"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractStorage-getItem}{}}}
\subsection{Method \code{getItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractStorage$getItem(key, ns = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractStorage-removeItem"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractStorage-removeItem}{}}}
\subsection{Method \code{removeItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractStorage$removeItem(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
