% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{input2storage}
\alias{input2storage}
\title{Save an input's property into browser's storage}
\usage{
input2storage(type, key, inputId, prop, session)
}
\arguments{
\item{type}{Kind of browser storage to use (localStorage ou sessionStorage)}

\item{key}{Identifier to store value}

\item{inputId}{The input's id}

\item{prop}{The input's prop to save}

\item{session}{Shiny session object}
}
\description{
Save an input's property directly in JavaScript into browser's storage
without reading the input from R.
}
