% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{LocalStorage}
\alias{LocalStorage}
\title{Binding of JS localStorage}
\description{
Binding of JS localStorage

Binding of JS localStorage
}
\section{Super class}{
\code{shinyBrowserStorage::AbstractStorage} -> \code{LocalStorage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalStorage-clone}{\code{LocalStorage$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="getItem"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-getItem'><code>shinyBrowserStorage::AbstractStorage$getItem()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="initialize"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-initialize'><code>shinyBrowserStorage::AbstractStorage$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="removeItem"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-removeItem'><code>shinyBrowserStorage::AbstractStorage$removeItem()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="setItem"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-setItem'><code>shinyBrowserStorage::AbstractStorage$setItem()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
