% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{read.storage}
\alias{read.storage}
\title{Read object from browser's localStorage or sessionStorage}
\usage{
read.storage(type, key, input, session, ns = NULL)
}
\arguments{
\item{type}{Kind of browser storage to use (localStorage ou sessionStorage)}

\item{key}{Identifier previoulsy used to store the wanted value}

\item{input}{Shiny input object}

\item{session}{Shiny session object}
}
\value{
a JSON decoded value (string, numeric, list, ...)
}
\description{
Read object from browser's localStorage or sessionStorage
}
