% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{remove.storage}
\alias{remove.storage}
\title{Remove value from browser's localStorage or sessionStorage}
\usage{
remove.storage(type, key, session)
}
\arguments{
\item{type}{Kind of browser storage to use (localStorage ou sessionStorage)}

\item{key}{Identifier previously used to store the wanted value}

\item{session}{Shiny session object}
}
\description{
Remove value from browser's localStorage or sessionStorage
}
