% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SessionStorage}
\alias{SessionStorage}
\title{Binding of JS sessionStorage}
\description{
Binding of JS sessionStorage

Binding of JS sessionStorage
}
\section{Super class}{
\code{shinyBrowserStorage::AbstractStorage} -> \code{SessionStorage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SessionStorage-clone}{\code{SessionStorage$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="getItem"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-getItem'><code>shinyBrowserStorage::AbstractStorage$getItem()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="initialize"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-initialize'><code>shinyBrowserStorage::AbstractStorage$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="removeItem"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-removeItem'><code>shinyBrowserStorage::AbstractStorage$removeItem()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyBrowserStorage" data-topic="AbstractStorage" data-id="setItem"><a href='../../shinyBrowserStorage/html/AbstractStorage.html#method-AbstractStorage-setItem'><code>shinyBrowserStorage::AbstractStorage$setItem()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SessionStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
