% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{withBrowserStorage}
\alias{withBrowserStorage}
\title{Define JS handlers required by shinyBrowserStorage's functions}
\usage{
withBrowserStorage()
}
\description{
Define JS handlers required by shinyBrowserStorage's functions
}
