% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write.storage}
\alias{write.storage}
\title{Write an object into browser's storage}
\usage{
write.storage(type, key, value, session)
}
\arguments{
\item{type}{Kind of browser storage to use (localStorage ou sessionStorage)}

\item{key}{Identifier to store value}

\item{value}{The value to be inserted in storage (need to be JSON encodable)}

\item{session}{Shiny session object}
}
\description{
Write an object into browser's storage
}
