% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localstorage.R
\name{LocalStorage}
\alias{LocalStorage}
\title{Binding of JS localStorage}
\description{
Binding of JS localStorage

Binding of JS localStorage
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalStorage-new}{\code{LocalStorage$new()}}
\item \href{#method-LocalStorage-setItem}{\code{LocalStorage$setItem()}}
\item \href{#method-LocalStorage-getItem}{\code{LocalStorage$getItem()}}
\item \href{#method-LocalStorage-removeItem}{\code{LocalStorage$removeItem()}}
\item \href{#method-LocalStorage-clone}{\code{LocalStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$new(input, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-setItem"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-setItem}{}}}
\subsection{Method \code{setItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$setItem(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-getItem"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-getItem}{}}}
\subsection{Method \code{getItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$getItem(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-removeItem"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-removeItem}{}}}
\subsection{Method \code{removeItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$removeItem(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
