% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionstorage.R
\name{SessionStorage}
\alias{SessionStorage}
\title{Binding of JS localStorage}
\description{
Binding of JS localStorage

Binding of JS localStorage
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SessionStorage-new}{\code{SessionStorage$new()}}
\item \href{#method-SessionStorage-setItem}{\code{SessionStorage$setItem()}}
\item \href{#method-SessionStorage-getItem}{\code{SessionStorage$getItem()}}
\item \href{#method-SessionStorage-removeItem}{\code{SessionStorage$removeItem()}}
\item \href{#method-SessionStorage-clone}{\code{SessionStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-SessionStorage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionStorage$new(input, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionStorage-setItem"></a>}}
\if{latex}{\out{\hypertarget{method-SessionStorage-setItem}{}}}
\subsection{Method \code{setItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionStorage$setItem(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionStorage-getItem"></a>}}
\if{latex}{\out{\hypertarget{method-SessionStorage-getItem}{}}}
\subsection{Method \code{getItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionStorage$getItem(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionStorage-removeItem"></a>}}
\if{latex}{\out{\hypertarget{method-SessionStorage-removeItem}{}}}
\subsection{Method \code{removeItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionStorage$removeItem(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SessionStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
